Calculadora :: Inserção da classe de Multiplicação e Divisão junto com os testes

+++ src/Operacoes/Soma.java
@@ -1,7 +1,7 @@
 package Operacoes;
 
 public class Soma {
-    public double Somar(double valor1, double valor2) {
+    public int Somar(int valor1, int valor2) {
         return valor1 + valor2;
     }
 }
