+++ src/Operacoes/Subtracao.java
@@ -1,7 +1,7 @@
 package Operacoes;
 
 public class Subtracao {
-    public double Subtrair(double valor1, double valor2) {
+    public int Subtrair(int valor1, int valor2) {
         return valor1 - valor2;
     }
 }
